Update main.cpp
add options to change display resolution and refresh_rate. Not quite as capable as gamestream launchpad but much faster to execute (no pyinstaller venv to decompress into memory) and works for newer games that play nice with borderless windowed mode. I also felt bad about not checking the CreateWindow return value so I now check to see if the window handle returned is null before trying to show or destroy the window.